Moved test route to folder
